requirements.txt
opencv-python>=4.8.0
numpy>=1.26.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pillow>=10.0.0
scikit-image>=0.22.0
pydantic>=2.5.0
requests>=2.31.0

